move event handlers into main handler
