--- package.json.orig
@@ -1,94 +1,0 @@
-{
-  "name": "hyperamp",
-  "productName": "Hyperamp",
-  "description": "Indescribable music player",
-  "version": "0.0.1",
-  "author": "Hyper Modules (https://github.com/hypermodules)",
-  "bugs": {
-    "url": "https://github.com/hypermodules/hyperamp/issues"
-  },
-  "contributors": [
-    "Bret Comnes <bcomnes@gmail.com> (http://bret.io)",
-    "Nate Goldman <ungoldman@gmail.com> (http://ungoldman.com)"
-  ],
-  "dependencies": {
-    "choo": "^4.1.0",
-    "choo-log": "^3.0.0",
-    "csjs-inject": "^1.0.1",
-    "electron-config": "^0.2.1",
-    "electron-context-menu": "^0.8.0",
-    "electron-debug": "^1.0.1",
-<<<<<<< HEAD
-    "electron-window-state": "^4.0.2",
-=======
-    "electron-default-menu": "^1.0.0",
->>>>>>> use electron-default-menu
-    "entypo": "^2.0.0",
-    "folder-walker": "^3.0.0",
-    "format-duration": "^1.0.0",
-    "musicmetadata": "^2.0.4"
-  },
-  "devDependencies": {
-    "@tap-format/spec": "^0.2.0",
-    "dependency-check": "^2.8.0",
-    "electron": "1.6.1",
-    "electron-builder": "^14.4.0",
-    "electron-rebuild": "^1.4.0",
-    "npm-run-all": "^4.0.1",
-    "snazzy": "^6.0.0",
-    "spectron": "^3.6.0",
-    "standard": "^9.0.0",
-    "tape": "^4.6.3",
-    "xvfb-maybe": "^0.1.3"
-  },
-  "homepage": "https://github.com/hypermodules/hyperamp",
-  "keywords": [
-    "amp",
-    "hyper",
-    "music",
-    "player"
-  ],
-  "license": "ISC",
-  "main": "main/index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/hypermodules/hyperamp.git"
-  },
-  "scripts": {
-    "start": "electron main",
-    "test": "run-s test:*",
-    "test:deps": "dependency-check ./package.json --entry renderer/index.js --no-dev --ignore-module electron",
-    "test:lint": "standard | snazzy",
-    "test:tape": "xvfb-maybe tape test/* | tap-format-spec",
-    "rebuild": "electron-rebuild",
-    "build": "build --dir",
-    "dist": "build"
-  },
-  "build": {
-    "appId": "com.hypermodules.hyperamp",
-    "mac": {
-      "category": "public.app-category.utilities",
-      "target": [
-        "zip",
-        "dmg"
-      ]
-    },
-    "nsis": {
-      "perMachine": true
-    },
-    "dmg": {
-      "contents": [
-        {
-          "x": 220,
-          "y": 200
-        },
-        {
-          "x": 448,
-          "y": 200,
-          "type": "link",
-          "path": "/Applications"
-        }
-      ]
-    }
-  }
-}
